requirements/prod.txt
Django==5.2.6
django_cleanup==9.0.0
django-environ==0.12.0
django-phonenumber-field[phonenumbers]==8.4.0
pillow==12.0.0
sorl-thumbnail==12.11.0